backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.7.1
google-generativeai==0.3.2
